bark-tts/requirements.txt
bark_ssg==1.3.4
scipy==1.14.0
